Performance and Helper functions additions
- This commit will not include recent changes from the official branch
since touch states are handled differently in this version than the
main branch after recent changes
- Added UIEnableScrollingForPanel, UIDestroyScroller, UISetupPaging,
UISetupPagingForPanel (see code for "how to use" for now)
- Call UISetupPaging in an immediate function
- Defined window.tapOrClick which tells the bind function to setup
either "tap" or "click" handlers
- Disabled bounce in iscroll by default to improve scroll performance
- Modified the inline style of the scrollbar in iscroll and removed
border radius to improve performance.
- Replaced use of document.querySelector with Zepto's querySelector in
chui.js
- Commented some console.log statements
- Added events for when a tab is displayed, hidden. Bind to
TabContent_Show and TabContent_Hide accordingly.
-Added event for when a detail view is loaded as a result of
clicking/tapping on a navigation list item. Bind to DetailViewEvent for
each of the list item to get called when this event is triggered.
- Modified UIBlock to show a CSS3 spinner by default and accepts a
display message and few other parameters (see code)
- Added loaders.css to support the CSS3 spinner that is display when
UIBlock is called (spinner can be hidden by passing "true" to UIBlock()
as the 4th parameter)
- Setting subviews to an intermediate class before changing them to
"upcoming" status from "current" status
- Zepto version 1.0rc1
- chui_overrides.css contains fix for tab views incorrectly overlapping
each other. This issue seemed to occur only in the Zepto version of
chui somehow.
- chui_overrides.css contains fix for hidden subview receiving tab
events or iOS' Next/Previous events and putting the app in an
inconsistent/broken state because it tries to focus on input boxes that
are in "upcoming" subviews